+++ packages/terra-clinical-header/tests/nightwatch/HeaderTestRoutes.jsx
@@ -10,6 +10,8 @@ import RightContentHeader from './RightContentHeader';
 import LeftContentHeader from './LeftContentHeader';
 import LeftAndRightContentHeader from './LeftAndRightContentHeader';
 import LongTitleAndContentHeader from './LongTitleAndContentHeader';
+import LeftAndRightContentSubheader from './LeftAndRightContentSubheader';
+import NoTitleRightContentHeader from './NoTitleRightContentHeader';
 
 const routes = (
   <div>
@@ -21,6 +23,8 @@ const routes = (
     <Route path="/tests/header-tests/left-content" component={LeftContentHeader} />
     <Route path="/tests/header-tests/left-and-right-content" component={LeftAndRightContentHeader} />
     <Route path="/tests/header-tests/long-title-content" component={LongTitleAndContentHeader} />
+    <Route path="/tests/header-tests/left-and-right-content-subheader" component={LeftAndRightContentSubheader} />
+    <Route path="/tests/header-tests/no-title-right-content" component={NoTitleRightContentHeader} />
   </div>
 );