+++ packages/terra-clinical-header/tests/nightwatch/HeaderTests.jsx
@@ -13,6 +13,8 @@ const HeaderTests = () => (
       <li><Link to="/tests/header-tests/left-content">Header - Left Content</Link></li>
       <li><Link to="/tests/header-tests/left-and-right-content">Header - Left and Right Content</Link></li>
       <li><Link to="/tests/header-tests/long-title-content">Header - Long Title and Content</Link></li>
+      <li><Link to="/tests/header-tests/left-and-right-content-subheader">Subheader - Left and Right Content</Link></li>
+      <li><Link to="/tests/header-tests/no-title-right-content">Header - No Title and RightContent</Link></li>
     </ul>
   </div>
 );