+++ packages/terra-clinical-header/tests/nightwatch/header-spec.js
@@ -32,4 +32,14 @@ module.exports = resizeTo(['tiny', 'small', 'medium', 'large', 'huge', 'enormous
     browser.url(`${browser.launchUrl}/#/tests/header-tests/long-title-content`)
     .assert.containsText('#Header h1', 'LongTitle');
   },
+  'Displays a subsection header with content on the left and right': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/header-tests/left-and-right-content-subheader`)
+    .assert.elementPresent('header[class*="flexSubheader"]');
+    browser.expect.element('#headerTest--startContent').to.be.present;
+    browser.expect.element('#headerTest--endContent').to.be.present;
+  },
+  'Displays a header with no title and content on the right': (browser) => {
+    browser.url(`${browser.launchUrl}/#/tests/header-tests/no-title-right-content`);
+    browser.expect.element('#headerTest--content').to.be.present;
+  },
 });