+++ packages/terra-clinical-detail-view/CHANGELOG.md
@@ -4,8 +4,12 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
-* Change data-class to be name spaced
 * Update nightwatch tests and test scripts.
+
+1.2.0 - (August 16, 2017)
+-----------------
+### Changed
+* Change data-class to be name spaced
 
 ### Fixes
 * Fixed divider rendering for when graphic is not present which caused double dividers