+++ packages/terra-clinical-header/CHANGELOG.md
@@ -3,11 +3,21 @@ ChangeLog
 
 Unreleased
 ----------
-### Fixed
-* Text Truncation in IE10 and IE11.
-
 ### Changed
 * Update nightwatch tests and test scripts.
+
+1.3.0 - (August 22, 2017)
+-----------------
+### Fixed
+* End aligned items with no title.
+
+### Added
+* isSubheader boolean prop to allow rendering of subheader component
+
+1.2.0 - (August 16, 2017)
+-----------------
+### Fixed
+* Text Truncation in IE10 and IE11.
 
 1.1.1 - (July 27, 2017)
 -----------------